refactor Node collections/card and create custom menu component, add delete option to menu
